Crate libbluec

The BlueC library.

Modules

codegen

The codegen module converts the BlueTac IR to an x86_64 Assembly AST and then writes/emits the assembly code to an ‘.s’ file. In the future, we’ll add other targets such as Arm64.

compiler_driver

The compiler_driver module defines the wrapper tool that orchestrates the different compilation stages.

ir

The ir module is responsible for lowering the C AST into the “BlueTac” intermediate representation (IR).

lexer

The lexer module handles the lexical analysis of the source code and then passes the results (a vector of tokens) to the parser.

parser

The parser module parses the stream of tokens produced by the lexer and produces an AST of the C code.

sema

The sema module is responsible for semantic analysis of the AST produced by the Parser.

Macros

ICE

Generates an internal compiler error.