libbluec

Module compiler_driver

The compiler_driver module defines the wrapper tool that orchestrates the different compilation stages.

Modules

args

The args module defines the command-line argument parser. Although crates like clap are excellent, we have a design goal to minimize dependencies on third-party crates and so BlueC uses a custom argument parser.

diagnostics

The diagnostics module defines errors and warnings emitted by the compiler driver.

errors

The errors module defines all the error diagnostics.

multi_file_driver

The multi_file_driver module provides functionality for multi-file compilation and linking.

options

The options module defines the compiler driver’s options.

tempfile

The tempfile module defines the TempFile type. This implementation is not as secure as the third-party tempfile crate but we have a design goal of no external dependencies for the BlueC library and executable.

warnings

The warnings module defines all the warning diagnostics.

Structs

Driver

The BlueC compiler driver.

Enums

CompilerGeneratedFile

A file path returned by the compiler driver.

DriverError

An error returned by the compiler driver.