libbluec::compiler_driver::options

Struct DriverOptions

pub struct DriverOptions {

    pub preprocessor_defns: Vec<String>,
    pub generate_object_file: bool,
    pub output_file: Option<String>,
    pub link_obj_files: Vec<String>,
    pub link_libs: Vec<String>,
    pub warnings_enabled: bool,
    pub warnings_as_errors: bool,
    pub only_create_asm_file: bool,
    pub lex: bool,
    pub parse: bool,
    pub validate: bool,
    pub codegen: bool,
    pub print_ast: bool,
    pub print_typechecked_ast: bool,
    pub print_ir: bool,
    pub flags: HashSet<String>,

}

Options that control the compiler driver’s behavior.

Fields
preprocessor_defns: Vec<String>

Define preprocessor macros.

generate_object_file: bool

Only run preprocess, compiler, and assembler steps to produce an object file (.o)

output_file: Option<String>

The file to write the output to, either an object file (.o) or an executable binary file.

link_obj_files: Vec<String>

Object files to link with.

link_libs: Vec<String>

Libraries to link with.

warnings_enabled: bool

Enable warnings.

warnings_as_errors: bool

Treat warnings as errors.

only_create_asm_file: bool

Only run preprocess and compiler steps to produce an assembly file (.s)

lex: bool

Run the lexer and then stop.

parse: bool

Run the lexer and parser and then stop.

validate: bool

Run the lexer, parser and semantic analysis and then stop.

codegen: bool

Run the lexer, parser, sema, IR translation, and assembly codegen and then stop.

print_ast: bool

Prints the parsed AST and stops after parsing (before sema).

print_typechecked_ast: bool

Prints the type-checked AST and stops after sema.

print_ir: bool

Prints the BlueTac intermediate representation and stops after lowering to IR.

flags: HashSet<String>

Flags which were passed on the command-line as ‘-f’.