libbluec::compiler_driver

Struct Driver

pub struct Driver {
    pub source_filename: String,
    pub translation_unit_filename: String,
    pub asm_filename: String,
    /* private fields */
}

The BlueC compiler driver.

Fields
source_filename: Stringtranslation_unit_filename: Stringasm_filename: String
Implementations

impl Driver

pub fn new(source_filename: &str, options: DriverOptions) -> Self

Creates a new compiler driver configured to compile the given source file.

pub fn run(&mut self) -> Result<CompilerGeneratedFile, DriverError>

Runs the compiler pipeline and returns the appropriate file depending on the given options.

pub fn options(&self) -> &DriverOptions

The compiler driver’s options.

pub fn take_options(&mut self) -> Option<DriverOptions>

Takes ownership of the compiler driver’s options.

pub fn is_flag_set(&self, flag: &str) -> bool

Is the given flag set in the options?

pub fn set_flag(&mut self, flag: &str)

Turns a flag on.

pub fn diagnostics_enabled(&self) -> bool

Are diagnostics enabled?

pub fn set_diagnostics_enabled(&mut self, enabled: bool)

Sets whether diagnostics are enabled.

pub fn add_diagnostic(&mut self, diagnostic: Diagnostic)

Adds a diagnostic (error or warning).

pub fn has_error_diagnostics(&self) -> bool

Are there any error diagnostics?

pub fn error_count(&self) -> usize

The number of error diagnostics.

pub fn warning_count(&self) -> usize

The number of warning diagnostics.

pub fn print_diagnostics(&self)

Prints all diagnostics to stderr, with any errors printed first before any warnings.

pub fn print_diagnostics_to_buffer(&self, buffer: impl Write)

Prints all diagnostics to the given buffer, with any errors printed first before any warnings.

You probably want print_diagnostics instead of this function, unless you deliberately want to print diagnostics into a buffer.

pub fn debug_print_diagnostics(&self)

For tests and debugging purposes, prints the diagnostics using the Debug trait.