libbluec::compiler_driver::warnings

Struct Warning

pub struct Warning;

Implementations

impl Warning

pub fn implicit_conversion( old_type: &AstType, new_type: &AstType, old_value: &str, new_value: &str, sign_change: bool, loc: SourceLocation, driver: &mut Driver, )

Emits an implicit conversion warning, optionally with a change in signedness.

pub fn implicit_switch_case_conversion( case_data_type: &AstType, switch_data_type: &AstType, old_value: &str, new_value: &str, loc: SourceLocation, driver: &mut Driver, )

Emits an implicit conversion warning for a switch statement case value.

pub fn extern_initializer(loc: SourceLocation, driver: &mut Driver)

Emits a warning about an ‘extern’ variable with an initializer.

pub fn duplicate_decl_specifier( specifier: SourceIdentifier<'_>, driver: &mut Driver, )

Emits a warning for a duplicate declaration specifier (e.g. duplicate storage class or ‘signed’/‘unsigned’).

-Wduplicate-decl-specifier

pub fn missing_declaration_identifier( is_typedef: bool, loc: SourceLocation, driver: &mut Driver, )

Emits a warning that a declaration is missing its identifier.

pub fn uninitialized_variable( variable: SourceIdentifier<'_>, driver: &mut Driver, )

Emits a warning that a variable is uninitialized.

pub fn integer_literal_implicitly_unsigned( loc: SourceLocation, driver: &mut Driver, )

Emits a warning that an integer literal is too large to be interpreted as signed and therefore is implicitly converted to unsigned.

pub fn unused_symbol( symbol: SourceIdentifier<'_>, kind: SymbolKind, driver: &mut Driver, )

Emits a warning that a symbol is unused.

-Wunused-variable, -Wunused-function, -Wunused-local-typedef

pub fn mixed_operators_missing_parens( child_expr_op: AstBinaryOp, parent_expr_op: AstBinaryOp, child_expr_loc: SourceLocation, parent_expr_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that an expression has mixed operators and is missing parentheses.

-Wlogical-op-parentheses, -Wbitwise-op-parentheses, -Wparentheses

pub fn assignment_in_condition_missing_parens( assignment_expr_loc: SourceLocation, assignment_operator_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that the result of an assignment is used as a condition without parentheses.

pub fn compare_different_pointer_types( a: &AstType, b: &AstType, cmp_loc: SourceLocation, a_loc: SourceLocation, b_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that two different pointer types are being compared.

-Wcompare-distinct-pointer-types

pub fn compare_pointer_and_integer( ptr_type: &AstType, int_type: &AstType, cmp_loc: SourceLocation, ptr_loc: SourceLocation, int_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that a pointer and an integer are being compared.

-Wpointer-integer-compare

pub fn pointer_to_smaller_int_cast( ptr_type: &AstType, int_type: &AstType, cmp_loc: SourceLocation, int_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that a pointer is casted to an integer type smaller than pointer type.

-Wpointer-to-int-cast

pub fn expression_interpreted_as_null_ptr_constant( loc: SourceLocation, ptr_type: &AstType, driver: &mut Driver, )

Emits a warning that an expression evaluated to zero and is interpreted as a null pointer constant.

-Wnon-literal-null-conversion

pub fn conditional_type_mismatch( a: &AstType, b: &AstType, ternary_loc: SourceLocation, a_loc: SourceLocation, b_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that a ternary condition’s consequent and alternative types do not match.

-Wconditional-type-mismatch

pub fn pointer_type_mismatch( a: &AstType, b: &AstType, expr_loc: SourceLocation, a_loc: SourceLocation, b_loc: SourceLocation, driver: &mut Driver, )

Emits a warning that two pointers in an expression have different types.

-Wpointer-type-mismatch