libbluec

Module ir

The ir module is responsible for lowering the C AST into the “BlueTac” intermediate representation (IR).

BlueTac is a custom three-address code (TAC) IR for the BlueC compiler. In the future, we’ll add a Static Single Assignment (SSA) constraint.

Structs

BtFunctionDefn

The IR for a function definition.

BtLabelIdentifier

A label identifier in IR.

BtRoot

The root of the BlueTac IR contains a list of definitions.

BtStaticStorageVariable

The IR for a variable with static storage duration.

BtSwitchCase

A case in a switch statement.

Enums

BtBinaryOp

Binary operator.

BtConstantValue

A constant integer or floating-point value.

BtDefinition

A BlueTac IR definition is either a function definition or a variable with static storage duration.

BtInstruction

A BlueTac IR instruction.

BtType

An IR data type.

BtUnaryOp

Unary operator.

BtValue

An IR value is either a constant or a variable.

Functions

translate

Translates the C AST produced by the parser into BlueTac intermediate representation (IR), and then passes the generated IR to the codegen stage.