libbluec::parser

Enum AstExpression

pub enum AstExpression {

    UnaryOperation {
        node_id: AstNodeId,
        op: AstUnaryOp,
        expr: Box<AstExpression>,
    },
    BinaryOperation {
        node_id: AstNodeId,
        op: AstBinaryOp,
        lhs: Box<AstExpression>,
        rhs: Box<AstExpression>,
    },
    Assignment {
        node_id: AstNodeId,
        computation_node_id: AstNodeId,
        op: AstAssignmentOp,
        lhs: Box<AstExpression>,
        rhs: Box<AstExpression>,
    },
    Conditional {
        node_id: AstNodeId,
        expr: Box<AstExpression>,
        consequent: Box<AstExpression>,
        alternative: Box<AstExpression>,
    },
    FunctionCall {
        node_id: AstNodeId,
        designator: Box<AstExpression>,
        args_node_id: AstNodeId,
        args: Vec<AstExpression>,
    },
    Deref {
        node_id: AstNodeId,
        expr: Box<AstExpression>,
    },
    AddressOf {
        node_id: AstNodeId,
        expr: Box<AstExpression>,
    },
    Subscript {
        node_id: AstNodeId,
        expr1: Box<AstExpression>,
        expr2: Box<AstExpression>,
    },
    Cast {
        node_id: AstNodeId,
        target_type: AstDeclaredType,
        expr: Box<AstExpression>,
        is_implicit: bool,
    },
    Identifier {
        node_id: AstNodeId,
        name: String,
        unique_name: AstUniqueName,
    },
    CharLiteral {
        node_id: AstNodeId,
        literal: String,
        value: i32,
    },
    StringLiteral {
        node_id: AstNodeId,
        literals: Vec<String>,
        ascii: Vec<String>,
    },
    IntegerLiteral {
        node_id: AstNodeId,
        literal: String,
        literal_base: usize,
        value: u64,
        kind: AstIntegerLiteralKind,
    },
    FloatLiteral {
        node_id: AstNodeId,
        literal: String,
        literal_base: usize,
        value: f64,
        kind: AstFloatLiteralKind,
    },

}

An expression, which may in fact be a subexpression inside a tree of a larger expression.

Variants

UnaryOperation

Fields

node_id: AstNodeId

op: AstUnaryOp

expr: Box<AstExpression>

BinaryOperation

Fields

node_id: AstNodeId

op: AstBinaryOp

lhs: Box<AstExpression>

rhs: Box<AstExpression>

Assignment

Fields

node_id: AstNodeId

computation_node_id: AstNodeId

op: AstAssignmentOp

lhs: Box<AstExpression>

rhs: Box<AstExpression>

Conditional

Fields

node_id: AstNodeId

expr: Box<AstExpression>

consequent: Box<AstExpression>

alternative: Box<AstExpression>

FunctionCall

Fields

node_id: AstNodeId

designator: Box<AstExpression>

args_node_id: AstNodeId

args: Vec<AstExpression>

Deref

Fields

node_id: AstNodeId

expr: Box<AstExpression>

AddressOf

Fields

node_id: AstNodeId

expr: Box<AstExpression>

Subscript

Fields

node_id: AstNodeId

expr1: Box<AstExpression>

expr2: Box<AstExpression>

Cast

Fields

node_id: AstNodeId

target_type: AstDeclaredType

expr: Box<AstExpression>

is_implicit: bool

Identifier

Fields

node_id: AstNodeId

name: String

unique_name: AstUniqueName

CharLiteral

Fields

node_id: AstNodeId

literal: String

value: i32

StringLiteral

Fields

node_id: AstNodeId

literals: Vec<String>

ascii: Vec<String>

IntegerLiteral

Fields

node_id: AstNodeId

literal: String

literal_base: usize

value: u64

kind: AstIntegerLiteralKind

FloatLiteral

Fields

node_id: AstNodeId

literal: String

literal_base: usize

value: f64

kind: AstFloatLiteralKind

Implementations

impl AstExpression

pub fn new_int_literal(value: u64) -> Self

Creates a new AstExpression:IntegerLiteral with the given value, in base 10.

pub fn node_id(&self) -> AstNodeId

Gets the node ID for the AST expression.

pub fn is_lvalue(&self) -> bool

Is the AST expression an l-value?

pub fn is_literal(&self) -> bool

Is the AST expression a literal?

pub fn is_arithmetic_literal(&self) -> bool

Is the AST expression an arithmetic (character, integer or floating-point) literal?

pub fn is_string_literal(&self) -> bool

Is the AST expression a string literal?

pub fn is_integer_literal(&self) -> bool

Is the AST expression an integer literal?

pub fn is_integer_literal_with_value(&self, value: u64) -> bool

Is the AST expression an integer literal with the given value?

pub fn is_identifier_with_name(&self, declared_name: &str) -> bool

Is the AST expression an identifier of the given declared name?

pub fn is_binary_expr(&self) -> bool

Is the AST expression a binary operation?

pub fn is_binary_expr_with_op(&self, op: AstBinaryOp) -> bool

Is the AST expression a binary operation with the given operator?