libbluec::parser

Enum AstType

pub enum AstType {

    Void,
    Short,
    Int,
    Long,
    LongLong,
    UnsignedShort,
    UnsignedInt,
    UnsignedLong,
    UnsignedLongLong,
    Float,
    Double,
    LongDouble,
    Pointer(Box<AstType>),
    Function {
        return_type: Box<AstType>,
        params: Vec<AstType>,
    },

}

The canonical type of an identifier.

Variants

Void

Short

Int

Long

LongLong

UnsignedShort

UnsignedInt

UnsignedLong

UnsignedLongLong

Float

Double

LongDouble

Pointer(Box<AstType>)

Function

Fields

return_type: Box<AstType>

params: Vec<AstType>

Implementations

impl AstType

pub fn new_pointer_to(ty: AstType) -> Self

Creates a new pointer type to the given AstType.

pub fn new_fn(return_type: AstType, params: Vec<AstType>) -> Self

Creates a new function type.

pub fn is_basic_type(&self) -> bool

Is this type a ‘basic type’?

The term ‘basic type’ is not used in the C Standard, but it is used informally when discussing C parsers. The basic type is the first type we see in a declaration before any of the declarators.

 unsigned long int *age = 0, calculate(float salary);
 ~~~~~~~~~~~~~~~~~

pub fn is_pointer(&self) -> bool

Is this type a pointer type?

pub fn is_function_pointer(&self) -> bool

Is this type a function pointer type?

A function pointer is a pointer whose referent is a function. In other words, there is only one level of indirection from the function.

In C, a function pointer can be called without needing to dereference it first.

int get(void);

int (*a)(void) = get;      // `a` is a function pointer.
int (**b)(void) = &a;      // `b` is _NOT_ a function pointer; it points to the object `a`.
 
int v1 = a();              // Can call `a` without dereferencing.
int v2 = (*a)();           // Or can explicitly dereference.

pub fn is_function(&self) -> bool

Is this type a function type? (Not a function pointer.)

pub fn is_arithmetic(&self) -> bool

Is this type an arithmetic type (integer, boolean, or floating-point)?

pub fn is_integer(&self) -> bool

Is this type an integral type?

pub fn is_floating_point(&self) -> bool

Is this type a floating-point type?

pub fn bits(&self) -> usize

The size of this type in bits.

pub fn is_signed_integer(&self) -> bool

Is this type a signed integer?

pub fn is_unsigned_integer(&self) -> bool

Is this type an unsigned integer?

pub fn same_signedness(&self, other: &AstType) -> bool

Does this type have the same signedness as the other type?

pub fn to_unsigned(self) -> Self

Makes an unsigned integer version of the current integer type.

Precondition: self.is_integer() == true

pub fn get_common_type(a: &AstType, b: &AstType) -> Self

Gets the common AstType for the given two types.

This does not perform any integer promotions, e.g. the common type for AstType::Short and AstType::Char is AstType::Short.

pub fn fits_inside(&self, other: &AstType) -> bool

Can all the values of this type fit inside the given other type?

Examples

assert_eq!(false, AstType::Int.fits_inside(&AstType::UnsignedInt));
assert_eq!(false, AstType::Int.fits_inside(&AstType::Short));
assert_eq!(true, AstType::Int.fits_inside(&AstType::Int));
assert_eq!(true, AstType::Int.fits_inside(&AstType::Long));
assert_eq!(false, AstType::UnsignedInt.fits_inside(&AstType::Int));
assert_eq!(true, AstType::UnsignedInt.fits_inside(&AstType::Long));
assert_eq!(true, AstType::Float.fits_inside(&AstType::Double));
assert_eq!(true, AstType::Double.fits_inside(&AstType::LongDouble));

pub fn can_hold_value(&self, value: u64) -> bool

Can the given u64 value fit inside the range of values supported by the current type?

Examples

assert_eq!(true, AstType::Short.can_hold_value(30_000_u64));
assert_eq!(false, AstType::Short.can_hold_value(100_000_u64));

pub fn promote_if_rank_lower_than_int(self) -> Self

Promotes a small integer type (‘_Bool’, ‘char’, or ‘short’) to AstType::Int.

If the type is an integral type with a rank lower than ‘int’ then the type is consumed and replaced with AstType::Int. Otherwise, the existing type is returned unchanged.

The C standard specifies that smaller integer types should be promoted to ‘unsigned int’ if their values cannot fit inside ‘int’, but in all mainstream implementations (including this one) all smaller integer types, both signed and unsigned, can fit into ‘int’.