libbluec::sema::constant_eval

Function evaluate_constant_full_expr

pub fn evaluate_constant_full_expr(
    full_expr: &AstFullExpression,
    ctx: ConstantEvalContext<'_, '_>,
) -> Option<AstConstantValue>

Evaluates an AST full expression node at compile-time and produces value, or None if it cannot be evaluated.

There is no undefined behaviour when evaluating a constant expression. If UB would have occurred then None is returned.