libbluec::sema

Module type_conversion

The type_conversion module provides functionality to convert between data types.

Functions

cast_i128_to_integer_type

Casts a i128 value to the given integer data type, and returns the new value as either u64 or i64.

convert_i32_to_i16

Converts an i32 value to an i16.

convert_i64_to_i16

Converts an i64 value to an i16.

convert_i64_to_i32

Converts an i64 value to an i32.

convert_u64_to_i8

Converts a u64 value to an i8.

convert_u64_to_i16

Converts a u64 value to an i16.

convert_u64_to_i32

Converts a u64 value to an i32.

convert_u64_to_i64

Converts a u64 value to an i64.

convert_u64_to_u8

Converts a u64 value to a u8.

convert_u64_to_u16

Converts a u64 value to a u16.

convert_u64_to_u32

Converts a u64 value to a u32.