libbluec

Module compiler_driver

The compiler_driver module defines the functions and types which orchestrate the different compilation stages.

Modules

args

The args module defines the command-line argument parser. Although crates like clap are excellent, we have a design goal to minimize dependencies on third-party crates and so BlueC uses a custom argument parser.

multi_file_driver

The multi_file_driver module provides functionality for multi-file compilation and linking.

options

The options module defines the compiler driver’s options.

Structs

Diagnostic

A diagnostic emitted by the compiler.

Driver

The BlueC compiler driver.

Error

An error diagnostic.

Warning

A warning diagnostic.

Enums

CompilerGeneratedFile

A file path returned by the compiler driver.

DriverError

An error returned by the compiler driver.

WarningKind

The kind of warning.